refactor(StyledComponents): Reuse wrapper border
Start abstracting out reusable pieces of CSS. This
is a good example as it was inconsistent across
components.
