+++ packages/react-component-library/src/styled-components/index.ts
@@ -1,0 +1,1 @@
+export * from './input'
